requirements.txt
Django==1.6.1
argparse==1.2.1
distribute==0.6.24
django-twilio==0.4
httplib2==0.8
six==1.5.2
twilio==3.6.4
unittest2==0.5.1
wsgiref==0.1.2

djangorestframework
markdown
django-filter
South
